+++ docs/sdk/go/examples/applications/app-queue/services/queue/model_catalog_queue_test.go
@@ -2,14 +2,15 @@ package queue
 
 import (
 	"fmt"
-	"github.com/hydraide/hydraide/docs/sdk/go/examples/applications/app-queue/utils/repo"
-	"github.com/hydraide/hydraide/sdk/go/hydraidego/client"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/suite"
 	"log/slog"
 	"os"
 	"testing"
 	"time"
+
+	"github.com/hydraide/hydraide/sdk/go/hydraidego/client"
+	"github.com/hydraide/hydraide/sdk/go/hydraidego/utils/repo"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/suite"
 )
 
 type TestModelCatalogQueue struct {
