+++ docs/sdk/go/examples/models/basics_subscribe.go
@@ -2,12 +2,13 @@ package models
 
 import (
 	"context"
-	"github.com/hydraide/hydraide/docs/sdk/go/examples/models/utils/hydraidehelper"
-	"github.com/hydraide/hydraide/docs/sdk/go/examples/models/utils/repo"
-	"github.com/hydraide/hydraide/sdk/go/hydraidego"
-	"github.com/hydraide/hydraide/sdk/go/hydraidego/name"
 	"log/slog"
 	"time"
+
+	"github.com/hydraide/hydraide/sdk/go/hydraidego"
+	"github.com/hydraide/hydraide/sdk/go/hydraidego/name"
+	"github.com/hydraide/hydraide/sdk/go/hydraidego/utils/hydraidehelper"
+	"github.com/hydraide/hydraide/sdk/go/hydraidego/utils/repo"
 )
 
 // ModelCatalogMessages demonstrates a simple socket-like message system powered by HydrAIDE.
