+++ docs/sdk/go/examples/models/catalog_save.go
@@ -1,11 +1,12 @@
 package models
 
 import (
-	"github.com/hydraide/hydraide/docs/sdk/go/examples/models/utils/hydraidehelper"
-	"github.com/hydraide/hydraide/docs/sdk/go/examples/models/utils/repo"
+	"time"
+
 	"github.com/hydraide/hydraide/sdk/go/hydraidego"
 	"github.com/hydraide/hydraide/sdk/go/hydraidego/name"
-	"time"
+	"github.com/hydraide/hydraide/sdk/go/hydraidego/utils/hydraidehelper"
+	"github.com/hydraide/hydraide/sdk/go/hydraidego/utils/repo"
 )
 
 // CatalogModelUserSaveExample defines a typed record (Treasure) inside a HydrAIDE Catalog Swamp.
