+++ docs/sdk/go/examples/models/model_catalog_example.go
@@ -1,6 +1,3 @@
-//go:build ignore
-// +build ignore
-
 // This file provides a detailed example of a catalog-style model used with CatalogCreate().
 // It explains required fields, supported types, optional metadata, and best practices.
